More changes to HasMQ
Make this role require an _mq param but override how this works in Crixa.pm so
it (and only it) will build a Net::AMQP::RabbitMQ object by default.

Also move the _connect_mq, connected, & disconnect methods so they're only
available from the Crixa package, not all packages with an mq object.