Allow non-Math::UInt64 delivery tags
The previous release changed the constraint on delivery_tag so
that a Math::UInt64 is always required. However,
Net::AMQP::RabbitMQ will pass a normal Perl scalar integer to
this module if the Perl install supports 64-bit unsigned
integers. This causes Crixa to throw an exception. This change
makes Crixa accept anything that looks like a positive integer.